Update houseplan_v00.py
houseplan_v00

Goal: A script for making a simple and fast plan (and free) for a place (a house, an appartment, etc.), using Turtle module. 

Description: The idea is subdivide the place into square rooms, measuring every one of them. Aside from the measures, we need to set the coordinates for every room; it will be considered that these coordinates start in the low-left corner (when looking to the screen). 
This data will be in tuples as: (x, y, distance x, distance y)
So, as an example for a room (0, 0, 372, 273): 
it means starts in (0, 0), distance x = 372 cm, distance y = 273 cm

How to use:
Between the lines #---prmts--- in the script, there is a real example (named "example: depto lazo"). This is also the place to set your own data for personal use. 
List "sbdvs" contains the dimensions of the rooms.
Dict "rooms" contains the names (real functions) of the rooms.

The main function "rects" is called at the end of the script.
Default mode contains:
- data for drawing
- height (default= 2.40): refers to height of the rooms (simplified as a common value)
- out_dims (default=True): displays the longest external dimensions
- intern_dims (default=True): displays every rooms' internal dimensions
- show_rooms (default=True): displays a list with rooms' names
- calc_data (default=True): it delivers the main measures of every room (useful for materials calculation).